+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/Compiler/UnshareAuthenticationResultHandlersPass.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\SecurityBundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Security\Http\Authentication\AuthenticationFailureHandlerInterface;
+use Symfony\Component\Security\Http\Authentication\AuthenticationSuccessHandlerInterface;
+
+class UnshareAuthenticationResultHandlersPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        foreach ($container->getDefinitions() as $definition) {
+            if (
+                is_a($definition->getClass(), AuthenticationFailureHandlerInterface::class, true)
+                || is_a($definition->getClass(), AuthenticationSuccessHandlerInterface::class, true)
+            ) {
+                $definition->setShared(false);
+            }
+        }
+    }
+}