+++ src/Symfony/Bundle/SecurityBundle/Tests/Functional/AuthenticatorTest.php
@@ -12,6 +12,10 @@
 namespace Symfony\Bundle\SecurityBundle\Tests\Functional;
 
 use PHPUnit\Framework\Attributes\DataProvider;
+use Symfony\Bundle\SecurityBundle\Tests\Fixtures\DecoratingAuthenticationFailureHandler;
+use Symfony\Bundle\SecurityBundle\Tests\Fixtures\DecoratingAuthenticationSuccessHandler;
+use Symfony\Component\Security\Http\Authentication\DefaultAuthenticationFailureHandler;
+use Symfony\Component\Security\Http\Authentication\DefaultAuthenticationSuccessHandler;
 
 class AuthenticatorTest extends AbstractWebTestCase
 {
@@ -93,12 +97,16 @@ class AuthenticatorTest extends AbstractWebTestCase
             '_password' => 'test',
         ]);
         $this->assertResponseRedirects('http://localhost/firewall1/test');
+        $this->assertResponseHeaderSame('X-Decorated-Handler', DefaultAuthenticationSuccessHandler::class);
+        $this->assertResponseHeaderSame('X-Decorating-Handler', DecoratingAuthenticationSuccessHandler::class);
 
         $client->request('POST', '/firewall1/dummy_login', [
             '_username' => 'jane@example.org',
             '_password' => 'test',
         ]);
         $this->assertResponseRedirects('http://localhost/firewall1/dummy');
+        $this->assertResponseHeaderSame('X-Decorated-Handler', DefaultAuthenticationSuccessHandler::class);
+        $this->assertResponseHeaderSame('X-Decorating-Handler', DecoratingAuthenticationSuccessHandler::class);
     }
 
     public function testCustomFailureHandler()
@@ -110,11 +118,15 @@ class AuthenticatorTest extends AbstractWebTestCase
             '_password' => 'wrong',
         ]);
         $this->assertResponseRedirects('http://localhost/firewall1/login');
+        $this->assertResponseHeaderSame('X-Decorated-Handler', DefaultAuthenticationFailureHandler::class);
+        $this->assertResponseHeaderSame('X-Decorating-Handler', DecoratingAuthenticationFailureHandler::class);
 
         $client->request('POST', '/firewall1/dummy_login', [
             '_username' => 'jane@example.org',
             '_password' => 'wrong',
         ]);
         $this->assertResponseRedirects('http://localhost/firewall1/dummy_login');
+        $this->assertResponseHeaderSame('X-Decorated-Handler', DefaultAuthenticationFailureHandler::class);
+        $this->assertResponseHeaderSame('X-Decorating-Handler', DecoratingAuthenticationFailureHandler::class);
     }
 }